# app/api/server.py
"""FastAPI boundary for the schema 5.0 multi-turn conversation product."""

from __future__ import annotations

import asyncio
import logging
import os
import secrets
from pathlib import Path
from typing import Any

from fastapi import (
    Depends,
    FastAPI,
    File,
    HTTPException,
    Request,
    Response,
    UploadFile,
    WebSocket,
)
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import FileResponse
from starlette.websockets import WebSocketDisconnect, WebSocketState

from app.api.events import ConversationEventBus
from app.api.schemas import (
    AdminUserResponse,
    AttachmentResponse,
    ConversationCreateRequest,
    ConversationRenameRequest,
    ConversationResponse,
    ConversationSummary,
    HealthResponse,
    LibraryDocument,
    ReadinessResponse,
    LoginRequest,
    LoginResponse,
    TurnResponse,
    TurnStartRequest,
    TurnStartResponse,
    UserResponse,
)
from app.conversation.application import ConversationApplication
from app.conversation.report import ConversationReport
from app.conversation.store import (
    Attachment,
    Conversation,
    ConversationStore,
    Turn,
    User,
)
from app.logging_setup import brief

_COOKIE = "deepsearch_session"
_LIBRARY_MAX_FILE_SIZE = 10 * 1024 * 1024  # 与 readers.MAX_FILE_SIZE_BYTES 一致
_DEFAULT_CORS_ORIGINS = (
    "http://127.0.0.1:5181",
    "http://localhost:5181",
    "http://127.0.0.1:5173",
    "http://localhost:5173",
)


def _cors_origins() -> list[str]:
    """DEEPSEARCH_CORS_ORIGINS 逗号分隔可覆写（H3；默认本地开发端口）。"""
    raw = os.getenv("DEEPSEARCH_CORS_ORIGINS", "").strip()
    if not raw:
        return list(_DEFAULT_CORS_ORIGINS)
    return [item.strip() for item in raw.split(",") if item.strip()]


def _cookie_secure() -> bool:
    return os.getenv("DEEPSEARCH_COOKIE_SECURE", "").strip().casefold() in {
        "1",
        "true",
        "yes",
        "on",
    }


def create_app(
    *,
    store: ConversationStore | None = None,
    conversation_application: ConversationApplication | Any | None = None,
    events: ConversationEventBus | None = None,
) -> FastAPI:
    """Build an isolated application; dependencies are injectable for tests."""

    if store is None:
        store = ConversationStore(
            Path(os.getenv("DEEPSEARCH_SQLITE", ".data/conversations.sqlite3"))
        )
    report = getattr(conversation_application, "report", None)
    if report is None:
        report = ConversationReport(
            Path(os.getenv("DEEPSEARCH_REPORT_ROOT", ".data/reports")), store
        )
    if events is None:
        events = ConversationEventBus()

    app = FastAPI(title="deepsearch-conversation", version="5.0.0")
    app.state.store = store
    app.state.conversation_application = conversation_application
    app.state.events = events
    # I5：task 注册表按 (conversation, turn) 键控，支持取消定位
    app.state.turn_tasks: dict[tuple[str, str], asyncio.Task[Any]] = {}
    app.add_middleware(
        CORSMiddleware,
        allow_origins=_cors_origins(),
        allow_credentials=True,
        allow_methods=["GET", "POST", "PATCH", "DELETE", "OPTIONS"],
        allow_headers=["Content-Type"],
    )

    def current_user(request: Request) -> User:
        user = store.resolve_session(request.cookies.get(_COOKIE))
        if user is None:
            raise HTTPException(status_code=401, detail="authentication required")
        return user

    def user_response(user: User) -> UserResponse:
        return UserResponse(id=user.id, username=user.username, role=user.role)

    def attachment_response(item: Attachment) -> AttachmentResponse:
        return AttachmentResponse(
            id=item.id,
            name=item.name,
            size=item.size,
            media_type=item.media_type,
            active=item.active,
        )

    def turn_response(item: Turn) -> TurnResponse:
        return TurnResponse(
            id=item.id,
            question=item.question,
            answer=item.answer,
            use_web=item.use_web,
            status=item.status,
            attachment_ids=list(item.attachment_ids),
            result=item.result,
            created_at=item.created_at,
            completed_at=item.completed_at,
        )

    def conversation_response(
        user: User, conversation: Conversation
    ) -> ConversationResponse:
        return ConversationResponse(
            id=conversation.id,
            title=conversation.title,
            owner_id=conversation.owner_id,
            created_at=conversation.created_at,
            updated_at=conversation.updated_at,
            turns=[
                turn_response(item) for item in store.list_turns(user, conversation.id)
            ],
            attachments=[
                attachment_response(item)
                for item in store.list_attachments(user, conversation.id)
            ],
        )

    def require_conversation(user: User, conversation_id: str) -> Conversation:
        try:
            return store.get_conversation(user, conversation_id)
        except LookupError as exc:
            raise HTTPException(
                status_code=404, detail="conversation not found"
            ) from exc

    @app.get("/ready", response_model=ReadinessResponse)
    async def ready() -> ReadinessResponse:
        # 3.8：readiness 反映真实研究能力，不执行真实 LLM 调用。
        # model 不可用 → not_ready；model 可用但证据源缺失 → degraded。
        caps = getattr(conversation_application, "capabilities", {}) or {}
        model_status = caps.get("model", {}).get("status", "unavailable")
        evidence_ready = [
            key
            for key in ("knowledge", "web")
            if caps.get(key, {}).get("status") == "ready"
        ]
        if model_status != "ready":
            status = "not_ready"
        elif evidence_ready:
            status = "ready"
        else:
            status = "degraded"
        return ReadinessResponse(
            status=status,
            capabilities=caps,
        )

    @app.get("/health", response_model=HealthResponse)
    async def health() -> HealthResponse:
        capabilities = getattr(
            conversation_application,
            "capabilities",
            {
                source: {"status": "unavailable"}
                for source in ("model", "knowledge", "web", "session_file")
            },
        )
        return HealthResponse(
            status="ok",
            capabilities=capabilities,
        )

    @app.post("/api/auth/login", response_model=LoginResponse)
    async def login(body: LoginRequest) -> LoginResponse:
        user = store.authenticate(body.username, body.password)
        if user is None:
            raise HTTPException(status_code=401, detail="invalid credentials")
        token = store.create_session(user)
        from fastapi.responses import JSONResponse

        result = JSONResponse(
            content=LoginResponse(user=user_response(user)).model_dump()
        )
        result.set_cookie(
            _COOKIE,
            token,
            httponly=True,
            secure=_cookie_secure(),
            samesite="lax",
            max_age=7 * 24 * 60 * 60,
        )
        return result  # type: ignore[return-value]

    @app.post("/api/auth/logout", status_code=204)
    async def logout(request: Request) -> None:
        token = request.cookies.get(_COOKIE)
        if token:
            store.delete_session(token)

    @app.get("/api/auth/me", response_model=UserResponse)
    async def me(user: User = Depends(current_user)) -> UserResponse:
        return user_response(user)

    @app.get("/api/admin/users", response_model=list[AdminUserResponse])
    async def admin_users(
        user: User = Depends(current_user),
    ) -> list[AdminUserResponse]:
        try:
            summaries = store.admin_user_summaries(user)
        except PermissionError as exc:
            raise HTTPException(
                status_code=403, detail="administrator access required"
            ) from exc
        return [
            AdminUserResponse(
                id=item.id,
                username=item.username,
                role=item.role,
                conversation_count=item.conversation_count,
            )
            for item in summaries
        ]

    @app.delete("/api/admin/users/{user_id}/data", status_code=204)
    async def admin_delete_user_data(
        user_id: str, user: User = Depends(current_user)
    ) -> None:
        try:
            # 会话行删除前先清外置资产：报告文件与个人知识库（G3 数据完整性）
            for conversation_id in store.admin_conversation_ids(user, user_id):
                report.discard(user, conversation_id)
            store.admin_delete_user_data(user, user_id)
        except PermissionError as exc:
            raise HTTPException(
                status_code=403, detail="administrator access required"
            ) from exc
        except ValueError as exc:
            raise HTTPException(
                status_code=422, detail="administrator data cannot be deleted here"
            ) from exc
        library = getattr(conversation_application, "upload_store", None)
        if library is not None:
            try:
                await asyncio.to_thread(library.delete_user, user_id)
            except Exception as exc:
                logging.getLogger("deepsearch.api").warning(
                    "cleanup uploads for user failed: %s", brief(exc)
                )

    @app.get("/api/conversations", response_model=list[ConversationResponse])
    async def list_conversations(
        user: User = Depends(current_user),
    ) -> list[ConversationResponse]:
        return [
            conversation_response(user, item) for item in store.list_conversations(user)
        ]

    @app.get("/api/conversations/lite", response_model=list[ConversationSummary])
    async def list_conversations_lite(
        user: User = Depends(current_user),
    ) -> list[ConversationSummary]:
        """轻量会话列表（G10）：只含元数据，供高频刷新场景使用。"""
        return [
            ConversationSummary(
                id=item.id,
                title=item.title,
                owner_id=item.owner_id,
                created_at=item.created_at,
                updated_at=item.updated_at,
            )
            for item in store.list_conversations(user)
        ]

    @app.post(
        "/api/conversations", response_model=ConversationResponse, status_code=201
    )
    async def create_conversation(
        body: ConversationCreateRequest,
        user: User = Depends(current_user),
    ) -> ConversationResponse:
        try:
            item = store.create_conversation(user, body.title)
        except ValueError as exc:
            raise HTTPException(
                status_code=422, detail="invalid conversation title"
            ) from exc
        return conversation_response(user, item)

    @app.get(
        "/api/conversations/{conversation_id}", response_model=ConversationResponse
    )
    async def get_conversation(
        conversation_id: str, user: User = Depends(current_user)
    ) -> ConversationResponse:
        return conversation_response(user, require_conversation(user, conversation_id))

    @app.patch(
        "/api/conversations/{conversation_id}", response_model=ConversationResponse
    )
    async def rename_conversation(
        conversation_id: str,
        body: ConversationRenameRequest,
        user: User = Depends(current_user),
    ) -> ConversationResponse:
        require_conversation(user, conversation_id)
        try:
            item = store.rename_conversation(user, conversation_id, body.title)
        except ValueError as exc:
            raise HTTPException(
                status_code=422, detail="invalid conversation title"
            ) from exc
        return conversation_response(user, item)

    @app.delete("/api/conversations/{conversation_id}", status_code=204)
    async def delete_conversation(
        conversation_id: str, user: User = Depends(current_user)
    ) -> None:
        require_conversation(user, conversation_id)
        report.discard(user, conversation_id)
        store.delete_conversation(user, conversation_id)

    @app.post(
        "/api/conversations/{conversation_id}/turns",
        response_model=TurnStartResponse,
        status_code=202,
    )
    async def start_turn(
        conversation_id: str,
        body: TurnStartRequest,
        user: User = Depends(current_user),
    ) -> TurnStartResponse:
        require_conversation(user, conversation_id)
        application = conversation_application
        if application is None:
            raise HTTPException(
                status_code=503, detail="conversation engine unavailable"
            )
        try:
            turn = await application.submit(
                user, conversation_id, question=body.question, use_web=body.use_web
            )
        except (LookupError, ValueError) as exc:
            raise HTTPException(status_code=422, detail="invalid turn") from exc
        events.emit(
            conversation_id, turn.id, "turn.started", "本轮分析已开始", stage="started"
        )
        task = asyncio.create_task(
            _execute_turn(application, user, conversation_id, turn.id, events)
        )
        app.state.turn_tasks[(conversation_id, turn.id)] = task
        task.add_done_callback(
            lambda _task, key=(conversation_id, turn.id): app.state.turn_tasks.pop(
                key, None
            )
        )
        return TurnStartResponse(
            turn_id=turn.id, conversation_id=conversation_id, use_web=turn.use_web
        )

    @app.get(
        "/api/conversations/{conversation_id}/turns/{turn_id}",
        response_model=TurnResponse,
    )
    async def get_turn(
        conversation_id: str,
        turn_id: str,
        user: User = Depends(current_user),
    ) -> TurnResponse:
        require_conversation(user, conversation_id)
        try:
            return turn_response(store.get_turn(user, conversation_id, turn_id))
        except LookupError as exc:
            raise HTTPException(status_code=404, detail="turn not found") from exc

    @app.delete("/api/conversations/{conversation_id}/turns/{turn_id}", status_code=204)
    async def cancel_turn(
        conversation_id: str,
        turn_id: str,
        user: User = Depends(current_user),
    ) -> Response:
        """I5：取消执行中的回合；僵尸 running（无任务）直接标记失败。"""
        require_conversation(user, conversation_id)
        try:
            current = store.get_turn(user, conversation_id, turn_id)
        except LookupError as exc:
            raise HTTPException(status_code=404, detail="turn not found") from exc
        if current.status != "running":
            raise HTTPException(status_code=409, detail="turn is already terminal")
        task = app.state.turn_tasks.get((conversation_id, turn_id))
        if task is not None:
            task.cancel()
        else:
            store.fail_turn(user, conversation_id, turn_id, "本轮研究已取消。")
            events.emit(
                conversation_id,
                turn_id,
                "turn.cancelled",
                "本轮研究已取消。",
                stage="failed",
                data={"turn_id": turn_id},
            )
        return Response(status_code=204)

    # 会话附件上传/列表/删除端点已随 T1 移除：附件功能由知识库入库
    # （个人知识库 collection）取代；历史 AttachmentResponse 字段与
    # store.attachments 表保留以维持读取兼容。

    @app.get("/api/conversations/{conversation_id}/report")
    async def download_report(
        conversation_id: str, user: User = Depends(current_user)
    ) -> FileResponse:
        require_conversation(user, conversation_id)
        path = store.report_path(user, conversation_id)
        if not path:
            raise HTTPException(status_code=404, detail="report not found")
        path = report.refresh(user, conversation_id)
        if not Path(path).is_file():
            raise HTTPException(status_code=404, detail="report not found")
        return FileResponse(
            path, media_type="text/markdown", filename="research-report.md"
        )

    # ---- 个人知识库（RAG 入库，T3）----

    def _library_store():
        library = getattr(conversation_application, "upload_store", None)
        if library is None:
            raise HTTPException(
                status_code=503, detail="personal knowledge base unavailable"
            )
        return library

    @app.get("/api/library/documents", response_model=list[LibraryDocument])
    async def list_library_documents(
        user: User = Depends(current_user),
    ) -> list[LibraryDocument]:
        return [
            LibraryDocument(
                document_id=item["document_id"],
                name=item["name"],
                chunks=int(item.get("chunks", "0") or 0),
            )
            for item in _library_store().list_documents(user.id)
        ]

    @app.post(
        "/api/library/documents",
        response_model=list[LibraryDocument],
        status_code=201,
    )
    async def ingest_library_documents(
        files: list[UploadFile] = File(...),
        user: User = Depends(current_user),
    ) -> list[LibraryDocument]:
        import tempfile

        library = _library_store()
        created: list[LibraryDocument] = []
        with tempfile.TemporaryDirectory(prefix="library-ingest-") as workdir:
            for upload in files:
                original = Path(upload.filename or "").name
                if not original or Path(original).suffix.lower() not in {
                    ".txt",
                    ".md",
                    ".pdf",
                    ".docx",
                    ".xlsx",
                }:
                    raise HTTPException(422, detail="不支持的文档类型")
                payload = await upload.read(_LIBRARY_MAX_FILE_SIZE + 1)
                if len(payload) > _LIBRARY_MAX_FILE_SIZE:
                    raise HTTPException(413, detail="文档过大")
                destination = Path(workdir) / f"{secrets.token_hex(8)}-{original}"
                destination.write_bytes(payload)
                try:
                    # 入库含嵌入计算与 Qdrant 写盘（秒级），转工作线程执行
                    # 避免阻塞事件循环（G8）
                    entry = await asyncio.to_thread(
                        library.ingest_path, user.id, original, destination
                    )
                except ValueError as exc:
                    raise HTTPException(422, detail=str(exc)) from exc
                except Exception as exc:
                    raise HTTPException(422, detail="文档解析失败") from exc
                created.append(
                    LibraryDocument(
                        document_id=entry["document_id"],
                        name=entry["name"],
                        chunks=int(entry.get("chunks", "0") or 0),
                    )
                )
        return created

    @app.delete("/api/library/documents/{document_id}", status_code=204)
    async def delete_library_document(
        document_id: str, user: User = Depends(current_user)
    ) -> Response:
        removed = await asyncio.to_thread(_library_store().remove, user.id, document_id)
        if not removed:
            raise HTTPException(status_code=404, detail="document not found")
        return Response(status_code=204)

    @app.websocket("/api/conversations/{conversation_id}/events")
    async def conversation_events(websocket: WebSocket, conversation_id: str) -> None:
        user = store.resolve_session(websocket.cookies.get(_COOKIE))
        if user is None:
            await websocket.close(code=4401)
            return
        try:
            store.get_conversation(user, conversation_id)
        except LookupError:
            await websocket.close(code=4404)
            return
        await websocket.accept()
        async with events.subscribe(conversation_id) as subscription:
            sender = asyncio.create_task(_send_events(websocket, subscription))
            overflow_watcher = asyncio.create_task(subscription.overflowed.wait())
            receive = asyncio.create_task(websocket.receive_text())
            try:
                while True:
                    done, _pending = await asyncio.wait(
                        {receive, overflow_watcher},
                        return_when=asyncio.FIRST_COMPLETED,
                    )
                    if overflow_watcher in done:
                        # 队列溢出后订阅已被丢弃，继续挂着只会静默失聪；
                        # 主动断开让前端走重连自愈（G11）
                        await websocket.close(code=1013)
                        break
                    if receive in done:
                        exc = receive.exception()
                        if exc is not None:
                            raise exc
                        receive = asyncio.create_task(websocket.receive_text())
            except WebSocketDisconnect:
                pass
            finally:
                sender.cancel()
                overflow_watcher.cancel()
                receive.cancel()

    return app


async def _execute_turn(
    application: Any,
    user: User,
    conversation_id: str,
    turn_id: str,
    events: ConversationEventBus,
) -> None:
    def emit(payload: dict[str, Any]) -> None:
        event_type = payload.get("type")
        if event_type not in {
            "stage.changed",
            "answer.delta",
            "evidence.ready",
            "report.updated",
            "turn.completed",
            "turn.failed",
            "turn.cancelled",
        }:
            return
        events.emit(
            conversation_id,
            turn_id,
            event_type,
            str(payload.get("message", "")),
            stage=payload.get("stage"),
            data=payload.get("data") or {},
        )

    await application.execute(user, conversation_id, turn_id, emit=emit)


async def _send_events(websocket: WebSocket, subscription: Any) -> None:
    while websocket.client_state == WebSocketState.CONNECTED:
        event = await subscription.queue.get()
        await websocket.send_text(event.model_dump_json())